+++ frontend/src/components/Guest View/GuestView.jsx
@@ -3,33 +3,12 @@ import { Chart as ChartJS, ArcElement, Tooltip, Legend } from "chart.js";
 import { Doughnut } from "react-chartjs-2";
 import "chart.js/auto";
 import DummyCalendar from "./DummyCalendar";
+import DummyChart from "./DummyChart";
 
 function GuestView() {
-  ChartJS.register(ArcElement, Tooltip, Legend);
-  const dataChart = {
-    labels: ["January", "February", "March", "April", "May"],
-    datasets: [
-      {
-        label: "Hours of Sleep",
-        backgroundColor: "rgba(75,192,192,0.2)",
-        borderColor: "rgba(75,192,192,1)",
-        borderWidth: 1,
-        hoverBackgroundColor: "rgba(75,192,192,0.4)",
-        hoverBorderColor: "rgba(75,192,192,1)",
-        data: [6, 7, 6.5, 7.5, 6.8],
-      },
-      {
-        backgroundColor: ["#FF5733", "#FFC300", "#C70039", "#900C3F"],
-      },
-    ],
-    options: {
-      responsive: true,
-    },
-  };
-
   return (
-    <div className="grid grid-cols-2 items-center justify-center p-10">
-      <div className="">
+    <div className="flex flex-col items-center justify-center pt-10">
+      <div className="flex flex-col w-1/2 justify-center p-3">
         <h1 className="font-heading">What is sleepyLog?</h1>
         <p>
           sleepyLog is a personalized sleep tracker app designed to help
@@ -38,11 +17,11 @@ function GuestView() {
           satisfy your curiosity and take action if there's any indicators
           through our data visualisation to having a good night's rest.
         </p>
-        <div className="calendar-container pb-2">
-          <DummyCalendar />
-        </div>
       </div>
-      <div>
+      <div className="calendar-container pb-2">
+        <DummyCalendar />
+      </div>
+      <div className="flex flex-col w-1/2 justify-center p-3">
         <h1 className="font-heading">Ok, great! But what do you do exactly?</h1>
         <p>
           Treat sleepyLog as an e-diary or sleep journal - but better! Unlike
@@ -53,13 +32,9 @@ function GuestView() {
           you've had a good night's rest within the past week, months, or even
           years!
         </p>
-        <div>
-          {Object.keys(dataChart).length > 0 ? (
-            <Doughnut data={dataChart} />
-          ) : (
-            <p>Loading...</p>
-          )}
-        </div>
+      </div>
+      <div className="pb-10 flex justify-center">
+        <DummyChart />
       </div>
     </div>
   );
